refactor: borderGradientStyle 파일명 변경
